requirements.txt
# Core dependencies
pillow==10.0.0
requests==2.31.0
configparser==5.3.0
caldav==1.4.0
pytz==2025.2
python-dotenv==1.0.0
